python_scripts/requirements.txt
kafka-python==2.0.2
watchdog==3.0.0
neo4j==5.14.1
torch==2.1.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
